+++ autosampler/spawners/__init__.py
@@ -4,3 +4,4 @@ from .density import DensitySpawner
 from .lof import LOFSpawner
 from .msm import MSMSpawner
 from .voronoi import VoronoiSpawner
+from .we import WESpawner